+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
-## 0.14 - 4/8/17
+## 0.14.1 - 5/13/17
+* `preserve_mode` actually does something now
+* added `debug(msg, persistent=False)` helper for scripts
+* bugfix where a fn decorated with `key_map` wasn't propagating additional options.
+
+## 0.14.0 - 4/8/17
 * Fixed `vim.buffers` indexing inconsistency between vim versions
 * Added polyfill for missing pyeval in older vim versions